# docs/swagger.yaml
definitions:
  dto.EventShortResponseDTO:
    properties:
      endDate:
        example: "2024-11-29"
        type: string
      endTime:
        example: "17:00"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: builds CMU
        type: string
      name:
        example: builds IDEA 2024
        type: string
      picUrl:
        example: https://example.com/image.jpg
        type: string
      startDate:
        example: "2024-11-29"
        type: string
      startTime:
        example: "08:00"
        type: string
    type: object
  dto.PaginatedEventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/dto.EventShortResponseDTO'
        type: array
      total_events:
        example: 1
        type: integer
    type: object
  fiber.Map:
    additionalProperties: true
    type: object
...
  handler.EventShortResponse:
    properties:
      endDate:
        type: string
      endTime:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      picUrl:
        type: string
      startDate:
        type: string
      startTime:
        type: string
    type: object
...
  models.Timeline:
    properties:
      activity:
        example: Registration
        type: string
      time:
        example: "08:00"
        type: string
    type: object
...
  service.EventCardResponses:
    properties:
      endDate:
        example: "2024-11-29"
        type: string
      endTime:
        example: "17:00"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: builds CMU
        type: string
      name:
        example: builds IDEA 2024
        type: string
      organization:
        $ref: '#/definitions/service.OrganizationShortRespones'
      organization_id:
        example: 1
        type: integer
      picUrl:
        example: https://example.com/image.jpg
        type: string
      province:
        example: Chiang Mai
        type: string
      startDate:
        example: "2024-11-29"
        type: string
      startTime:
        example: "08:00"
        type: string
    type: object
  service.EventResponses:
    properties:
      description:
        example: This is a description
        type: string
      endDate:
        example: "2024-11-29"
        type: string
      endTime:
        example: "17:00"
        type: string
      highlight:
        example: This is a highlight
        type: string
      id:
        example: 1
        type: integer
      keyTakeaway:
        example: This is a key takeaway
        type: string
      latitude:
        example: "13.7563"
        type: string
      locationName:
        example: builds CMU
        type: string
      longitude:
        example: "100.5018"
        type: string
      name:
        example: builds IDEA 2024
        type: string
      organization_id:
        example: 1
        type: integer
      picUrl:
        example: https://example.com/image.jpg
        type: string
      province:
        example: Chiang Mai
        type: string
      requirement:
        example: This is a requirement
        type: string
      startDate:
        example: "2024-11-29"
        type: string
      startTime:
        example: "08:00"
        type: string
      timeLine:
        items:
          $ref: '#/definitions/models.Timeline'
        type: array
    type: object
  service.NewEventRequest:
    properties:
      description:
        example: This is a description
        type: string
      endDate:
        example: "2024-11-29"
        type: string
      endTime:
        example: "17:00"
        type: string
      highlight:
        example: This is a highlight
        type: string
      keyTakeaway:
        example: This is a key takeaway
        type: string
      latitude:
        example: "13.7563"
        type: string
      locationName:
        example: Bangkok
        type: string
      longitude:
        example: "100.5018"
        type: string
      name:
        example: builds IDEA 2024
        type: string
      picUrl:
        example: https://example.com/image.jpg
        type: string
      province:
        example: Chiang Mai
        type: string
      requirement:
        example: This is a requirement
        type: string
      startDate:
        example: "2024-11-29"
        type: string
      startTime:
        example: "08:00"
        type: string
      timeLine:
        items:
          $ref: '#/definitions/models.Timeline'
        type: array
    type: object
  service.OrganizationShortRespones:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: builds CMU
        type: string
      picUrl:
        example: https://example.com/image.jpg
        type: string
    type: object
info:
  contact: {}
paths:
  /events:
    get:
      description: Get a list of all events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/service.EventResponses'
              type: array
            type: array
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: List all events
      tags:
      - Events
  /events-paginate:
    get:
      description: Get a paginated list of events
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedEventsResponse'
        "400":
          description: Bad Request - Invalid page number
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Paginate events
      tags:
      - Events
  /events/first:
    get:
      description: Get the first event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.EventShortResponse'
            type: array
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get the first event
      tags:
      - Events
  /events/q:
    get:
      consumes:
      - application/json
      description: Search for mock events based on various criteria such as page,
        search term, category, location type, audience, and price type.
      parameters:
      - description: Page number
        in: query
        name: page
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Event category
        in: query
        name: category
        type: string
      - description: Location type
        in: query
        name: location
        type: string
      - description: Audience type
        in: query
        name: audience
        type: string
      - description: Price type
        in: query
        name: price
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.EventCardResponses'
            type: array
        "400":
          description: Bad Request - Invalid query parameters
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found - Events not found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Search for mock events
      tags:
      - Events
  /events/upcoming:
    get:
      description: Get a list of upcoming events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.EventResponses'
            type: array
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get upcoming events
      tags:
      - Events
...
  /orgs/{orgID}/events:
    get:
      description: Get a list of all events for a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/service.EventResponses'
              type: array
            type: array
        "400":
          description: Bad Request - Invalid organization id or missing orgID parameters
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: List all events for a specific organization
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event for a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/service.NewEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.EventResponses'
        "400":
          description: Bad Request - organization id is required or invalid / event
            body is invalid
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create a new event
      tags:
      - Events
  /orgs/{orgID}/events/{id}:
    delete:
      description: Delete an event by its ID for a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.EventResponses'
        "400":
          description: Bad Request - Invalid organization id or missing orgID parameters
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete an event
      tags:
      - Events
    get:
      description: Get an event by its ID for a specific organization
      parameters:
      - description: Organization ID
        in: path
        name: orgID
        required: true
        type: integer
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.EventResponses'
            type: array
        "400":
          description: Bad Request - Invalid organization id or missing orgID parameters
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error - Something went wrong
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get an event by ID
      tags:
      - Events
...
swagger: "2.0"